+++ packages/gql.client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gymlabs/core.gql.client
 
+## 0.1.4
+
+### Patch Changes
+
+- workout, exercise crud - fix types
+
+## 0.1.3
+
+### Patch Changes
+
+- add create, update, read for gyms
+
+## 0.1.2
+
+### Patch Changes
+
+- only delete user when it has no memberships or employments and reactivate account
+
 ## 0.1.1
 
 ### Patch Changes